!!! TASK: Cleanup in package management
This is the next step towards a leaner package management,
the essential part is that packages are now separated into
third party packages and Flow (enabled) packages.
All packages are available for object management but Resources
and Configuration as well as booting are only expected and
managed in Flow (enabled) packages.

The ``Package`` class is still a fully Flow enabled package and
no adaption should be necessary to packages.
``GenericPackage`` is the low level class for describing any
other package in the system.
According to that change a couple of interfaces where added:

* ``BootableInterface`` describes a bootable package
* ``FlowPackageInterface`` extension of the ``PackageInterface``
  Flow specifics are now moved over to the ``FlowPackageInterface``
* ``PackageKeyAwareInterface`` defines that the package has a
  package key. Currently that is implemented by all packages but
  we might change that at a later point in time.

The notion of ``protected`` and ``objectManagementEnabled`` is gone from
package classes and the interfaces as both are no longer needed.

The ``PackageManager`` and interface no longer support deleting of
packages, this should happen through composer now.
In order the ``package:delete`` command is removed as well.

``PackageManager::getPackageVersion`` was moved over to
``ComposerUtility::getPackageVersion`` where it should have been in the
first place. It was not part of the interface nor marked api.

This change is breaking if you use the ``PackageManager`` to get
all packages and expect them to be Flow packages. You must now
check for instanceof ``FlowPackageInterface`` if you expect Flow
specific functionality from a package object.

This is also a preparation to drop the ``PackageManagerInterface`` as
overriding the implementation is neither possible nor sensible.
You can directly use the ``PackageManager`` object from now.
