+++ Neos.Flow/Classes/Configuration/ConfigurationManager.php
@@ -13,7 +13,7 @@ namespace Neos\Flow\Configuration;
 
 use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Core\ApplicationContext;
-use Neos\Flow\Package\PackageInterface;
+use Neos\Flow\Package\FlowPackageInterface;
 use Neos\Utility\Arrays;
 use Neos\Utility\Files;
 use Neos\Utility\OpcodeCacheHelper;
@@ -165,7 +165,7 @@ class ConfigurationManager
     /**
      * Active packages to load the configuration for
      *
-     * @var array<Neos\Flow\Package\PackageInterface>
+     * @var FlowPackageInterface[]
      */
     protected $packages = [];
 
@@ -234,7 +234,7 @@ class ConfigurationManager
     /**
      * Sets the active packages to load the configuration for
      *
-     * @param array<Neos\Flow\Package\PackageInterface> $packages
+     * @param FlowPackageInterface[] $packages
      * @return void
      */
     public function setPackages(array $packages)
@@ -416,7 +416,7 @@ class ConfigurationManager
      * getConfiguration() method.
      *
      * @param string $configurationType The kind of configuration to load - must be one of the CONFIGURATION_TYPE_* constants
-     * @param array $packages An array of Package objects (indexed by package key) to consider
+     * @param FlowPackageInterface[] $packages An array of Package objects (indexed by package key) to consider
      * @throws Exception\InvalidConfigurationTypeException
      * @throws Exception\InvalidConfigurationException
      * @return void
@@ -440,7 +440,6 @@ class ConfigurationManager
                 }
 
                 $settings = [];
-                /** @var $package PackageInterface */
                 foreach ($packages as $packageKey => $package) {
                     if (Arrays::getValueByPath($settings, $packageKey) === null) {
                         $settings = Arrays::setValueByPath($settings, $packageKey, []);
@@ -450,7 +449,6 @@ class ConfigurationManager
                 $settings = Arrays::arrayMergeRecursiveOverrule($settings, $this->configurationSource->load(FLOW_PATH_CONFIGURATION . $configurationType, $allowSplitSource));
 
                 foreach ($this->orderedListOfContextNames as $contextName) {
-                    /** @var $package PackageInterface */
                     foreach ($packages as $package) {
                         $settings = Arrays::arrayMergeRecursiveOverrule($settings, $this->configurationSource->load($package->getConfigurationPath() . $contextName . '/' . $configurationType, $allowSplitSource));
                     }
@@ -466,7 +464,6 @@ class ConfigurationManager
                 $this->configurations[$configurationType]['Neos']['Flow']['core']['context'] = (string)$this->context;
             break;
             case self::CONFIGURATION_PROCESSING_TYPE_OBJECTS:
-                /** @var $package PackageInterface */
                 foreach ($packages as $packageKey => $package) {
                     $configuration = $this->configurationSource->load($package->getConfigurationPath() . $configurationType);
                     $configuration = Arrays::arrayMergeRecursiveOverrule($configuration, $this->configurationSource->load(FLOW_PATH_CONFIGURATION . $configurationType));
@@ -487,7 +484,6 @@ class ConfigurationManager
                     }
                 }
 
-                /** @var $package PackageInterface */
                 foreach ($packages as $package) {
                     $packagePolicyConfiguration = $this->configurationSource->load($package->getConfigurationPath() . $configurationType, $allowSplitSource);
                     $this->configurations[$configurationType] = $this->mergePolicyConfiguration($this->configurations[$configurationType], $packagePolicyConfiguration);
@@ -495,7 +491,6 @@ class ConfigurationManager
                 $this->configurations[$configurationType] = $this->mergePolicyConfiguration($this->configurations[$configurationType], $this->configurationSource->load(FLOW_PATH_CONFIGURATION . $configurationType, $allowSplitSource));
 
                 foreach ($this->orderedListOfContextNames as $contextName) {
-                    /** @var $package PackageInterface */
                     foreach ($packages as $package) {
                         $packagePolicyConfiguration = $this->configurationSource->load($package->getConfigurationPath() . $contextName . '/' . $configurationType, $allowSplitSource);
                         $this->configurations[$configurationType] = $this->mergePolicyConfiguration($this->configurations[$configurationType], $packagePolicyConfiguration);
@@ -504,14 +499,12 @@ class ConfigurationManager
                 }
             break;
             case self::CONFIGURATION_PROCESSING_TYPE_DEFAULT:
-                /** @var $package PackageInterface */
                 foreach ($packages as $package) {
                     $this->configurations[$configurationType] = Arrays::arrayMergeRecursiveOverrule($this->configurations[$configurationType], $this->configurationSource->load($package->getConfigurationPath() . $configurationType, $allowSplitSource));
                 }
                 $this->configurations[$configurationType] = Arrays::arrayMergeRecursiveOverrule($this->configurations[$configurationType], $this->configurationSource->load(FLOW_PATH_CONFIGURATION . $configurationType, $allowSplitSource));
 
                 foreach ($this->orderedListOfContextNames as $contextName) {
-                    /** @var $package PackageInterface */
                     foreach ($packages as $package) {
                         $this->configurations[$configurationType] = Arrays::arrayMergeRecursiveOverrule($this->configurations[$configurationType], $this->configurationSource->load($package->getConfigurationPath() . $contextName . '/' . $configurationType, $allowSplitSource));
                     }
@@ -530,7 +523,6 @@ class ConfigurationManager
                 $this->mergeRoutesWithSubRoutes($this->configurations[$configurationType]);
             break;
             case self::CONFIGURATION_PROCESSING_TYPE_APPEND:
-                /** @var $package PackageInterface */
                 foreach ($packages as $package) {
                     $this->configurations[$configurationType] = array_merge($this->configurations[$configurationType], $this->configurationSource->load($package->getConfigurationPath() . $configurationType, $allowSplitSource));
                 }
@@ -706,7 +698,6 @@ class ConfigurationManager
                 if (!isset($this->packages[$subRouteOptions['package']])) {
                     throw new Exception\ParseErrorException(sprintf('The SubRoute Package "%s" referenced in Route "%s" is not available.', $subRouteOptions['package'], (isset($routeConfiguration['name']) ? $routeConfiguration['name'] : 'unnamed Route')), 1318414040);
                 }
-                /** @var $package PackageInterface */
                 $package = $this->packages[$subRouteOptions['package']];
                 $subRouteFilename = 'Routes';
                 if (isset($subRouteOptions['suffix'])) {
