+++ Neos.Flow/Classes/Configuration/ConfigurationSchemaValidator.php
@@ -44,7 +44,7 @@ class ConfigurationSchemaValidator
 
     /**
      * @Flow\Inject
-     * @var \Neos\Flow\Package\PackageManagerInterface
+     * @var \Neos\Flow\Package\PackageManager
      */
     protected $packageManager;
 
@@ -99,8 +99,7 @@ class ConfigurationSchemaValidator
 
         // find schema files for the given type and path
         $schemaFileInfos = [];
-        $activePackages = $this->packageManager->getAvailablePackages();
-        foreach ($activePackages as $package) {
+        foreach ($this->packageManager->getFlowPackages() as $package) {
             $packageKey = $package->getPackageKey();
             $packageSchemaPath = Files::concatenatePaths([$package->getResourcesPath(), 'Private/Schema']);
             if (is_dir($packageSchemaPath)) {
