+++ Neos.Flow/Classes/I18n/Xliff/Service/XliffFileProvider.php
@@ -18,7 +18,7 @@ use Neos\Flow\I18n\Locale;
 use Neos\Flow\I18n\Xliff\Model\FileAdapter;
 use Neos\Flow\I18n\Xliff\V12\XliffParser as V12XliffParser;
 use Neos\Flow\Package\PackageInterface;
-use Neos\Flow\Package\PackageManagerInterface;
+use Neos\Flow\Package\PackageManager;
 use Neos\Utility\Arrays;
 use Neos\Utility\Files;
 
@@ -31,7 +31,7 @@ class XliffFileProvider
 {
     /**
      * @Flow\Inject
-     * @var PackageManagerInterface
+     * @var PackageManager
      */
     protected $packageManager;
 
@@ -101,8 +101,7 @@ class XliffFileProvider
             $localeChain = $this->localizationService->getLocaleChain($locale);
             // Walk locale chain in reverse, so that translations higher in the chain overwrite fallback translations
             foreach (array_reverse($localeChain) as $localeChainItem) {
-                foreach ($this->packageManager->getAvailablePackages() as $package) {
-                    /** @var PackageInterface $package */
+                foreach ($this->packageManager->getFlowPackages() as $package) {
                     $translationPath = $package->getResourcesPath() . $this->xliffBasePath . $localeChainItem;
                     if (is_dir($translationPath)) {
                         $this->readDirectoryRecursively($translationPath, $parsedData, $fileId, $package->getPackageKey());
