+++ Neos.Flow/Classes/Package/PackageInterface.php
@@ -10,32 +10,16 @@ namespace Neos\Flow\Package;
  * information, please view the LICENSE file which was distributed with this
  * source code.
  */
-use Neos\Flow\Core\Bootstrap;
 
 /**
- * Interface for a Flow Package class
+ * Interface for a basic Package class
  *
  * @api
  */
 interface PackageInterface
 {
     const PATTERN_MATCH_PACKAGEKEY = '/^[a-z0-9]+\.(?:[a-z0-9][\.a-z0-9]*)+$/i';
-
-    const DIRECTORY_CLASSES = 'Classes/';
-    const DIRECTORY_CONFIGURATION = 'Configuration/';
-    const DIRECTORY_TESTS_FUNCTIONAL = 'Tests/Functional/';
-    const DIRECTORY_TESTS_UNIT = 'Tests/Unit/';
-    const DIRECTORY_RESOURCES = 'Resources/';
-
     const DEFAULT_COMPOSER_TYPE = 'neos-package';
-
-    /**
-     * Invokes custom PHP code directly after the package manager has been initialized.
-     *
-     * @param Bootstrap $bootstrap The current bootstrap
-     * @return void
-     */
-    public function boot(Bootstrap $bootstrap);
 
     /**
      * Returns the array of filenames of the class files
@@ -44,14 +28,6 @@ interface PackageInterface
      * @api
      */
     public function getClassFiles();
-
-    /**
-     * Returns the package key of this package.
-     *
-     * @return string
-     * @api
-     */
-    public function getPackageKey();
 
     /**
      * Returns the composer name of this package.
@@ -70,30 +46,6 @@ interface PackageInterface
     public function getNamespaces();
 
     /**
-     * Tells if this package is protected and therefore cannot be deactivated or deleted
-     *
-     * @return boolean
-     * @api
-     */
-    public function isProtected();
-
-    /**
-     * Tells if files in the Classes directory should be registered and object management enabled for this package.
-     *
-     * @return boolean
-     */
-    public function isObjectManagementEnabled();
-
-    /**
-     * Sets the protection flag of the package
-     *
-     * @param boolean $protected TRUE if the package should be protected, otherwise FALSE
-     * @return void
-     * @api
-     */
-    public function setProtected($protected);
-
-    /**
      * Returns the full path to this package's main directory
      *
      * @return string Path to this package's main directory
@@ -101,21 +53,6 @@ interface PackageInterface
      */
     public function getPackagePath();
 
-    /**
-     * Returns the full path to this package's Resources directory
-     *
-     * @return string Path to this package's Resources directory
-     * @api
-     */
-    public function getResourcesPath();
-
-    /**
-     * Returns the full path to this package's Configuration directory
-     *
-     * @return string Path to this package's Configuration directory
-     * @api
-     */
-    public function getConfigurationPath();
 
     /**
      * Returns the currently installed version of this package.
