+++ Neos.Flow/Configuration/Objects.yaml
@@ -430,3 +430,13 @@ Neos\Flow\Session\SessionManager:
         arguments:
           1:
             value: Flow_Session_MetaData
+
+#
+# Utility
+#
+
+Neos\Utility\SchemaGenerator:
+  scope: singleton
+
+Neos\Utility\SchemaValidator:
+  scope: singleton
