--- Neos.Utility.Schema/Configuration/Objects.yaml
@@ -1,5 +1,0 @@
-Neos\Utility\SchemaGenerator:
-  scope: singleton
-
-Neos\Utility\SchemaValidator:
-  scope: singleton
